v3.0.0 first pass
