+++ api/draw.md
@@ -1,0 +1,41 @@
+<a name="api-sg-functions-draw-data"></a>
+## Draw Data Functions ##
+
+<a name="api-sg-drawPoints"></a>
+**`SimpleGraph.drawPoints()`** : (Re)draw all points data on graph. Points will have class `.sg-point`.
+
+&nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
+
+<a name="api-sg-drawLines"></a>
+**`SimpleGraph.drawLines()`** : (Re)draw all points data on graph. Lines will have class `.sg-line` or `.sg-plotted-line`.
+
+&nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
+
+<a name="api-sg-drawAreas"></a>
+**`SimpleGraph.drawAreas()`** : (Re)draw all area data on graph. Areas will have class `.plotted-areas`.
+
+&nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
+
+
+<a name="api-sg-functions-remove-graph"></a>
+## Remove From Graph Functions ##
+
+<a name="api-sg-removePoints"></a>
+**`SimpleGraph.removePoints()`** : Remove all drawn points on graph. Does not remove the underlying data.
+
+&nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
+
+<a name="api-sg-removeLines"></a>
+**`SimpleGraph.removeLines()`** : Remove all drawn lines on graph. Does not remove the underlying data.
+
+&nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
+
+<a name="api-sg-removeAreas"></a>
+**`SimpleGraph.removeAreas()`** : Remove all drawn areas on graph. Does not remove the underlying data.
+
+&nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
+
+<a name="api-sg-removeAll"></a>
+**`SimpleGraph.removeAll()`** : Remove all drawn data series on graph. Does not remove the underlying data.
+
+&nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
