+++ src/entry.js
@@ -1,0 +1,4 @@
+import polyfills from './polyfills';
+import SimpleGraph from './simple-graph.js';
+
+export default SimpleGraph;
