+++ tutorials/simplegraph-2.html
@@ -3,7 +3,7 @@
   <head>
     <title>SimpleGraph Additional Functionality</title>
     <meta charset="UTF-8">
-    <script src="https://d3js.org/d3.v4.min.js"></script>
+    <script src="https://d3js.org/d3.v6.min.js"></script>
     <script src="../d3.simplegraph.min.js"></script>
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/2.2.0/jquery.min.js"></script>
     <style>
